Update patch set 2
Patch Set 2:

The use of punctuation is somewhat ambiguous in this case (half width VS full width). For CM, we've been using half width in the past and for quite some time now, and even Google uses both of them rather inconsistently throughout the AOSP strings.

This is why I only committed to the CM fork: Some of the things were only done for consistency and may not necessarily apply to the Superuser Koush is maintaining.

Patch-set: 2