Update patch set 2
Patch Set 2:

huawuxin has been notified.

I do not intent/want Superuser in CM and my market release to be complete forks-- especially due to a translation issue.

AOSP practices are guidelines, and not absolute, as we are not AOSP (especially when they themselves are inconsistent). We should focus on what provides the best user experience- and if that means providing different Chinese locales, so be it.

I reverted your commit, Andrew, because it seemed quite aggressive in that it deleted 3 other translations. Granted, I do not understand Chinese, it seemed a little suspect, especially when the original author(s) were not consulted (and were not happy with the changes).

Let's work with huawuxin to come to a solution that works for everyone.

Patch-set: 2