Update patch set 2
Patch Set 2:

I think this might make sense for 10.2.

We scaled it back in previous versions, I think it's safe to disable it by default now and require the user to turn it on. I don't want to get rid of the ability to use root, but I also don't like it being on at all on first boot.

We might want to show a toast or something if su is denied.

Patch-set: 2