Update patch set 2
Patch Set 2:

If we can come to a consensus on this I will test if this is the way we want to go.  I am familiar with this issue as I have to let recovery update SU everytime I flash CM for root to work. SU fails to update with the system running.

I will test if this is the route we want to go.  If not then abandon?

Patch-set: 2