Update patch set 2
Patch Set 2:

_If_ you wanted to do this "start by default" approach (which you don't), you'd need to class it, yes.

I told you to remove it because we _don't_ want to start it automagically by default. We want to start it as a reaction to the property's value, and for that we should set a default value for said property. It won't be the first or only subsystem with this kind of requirement (usb composition being the most notable example. If a persist.sys.usb.config isn't part of the default properties, most devices will have dead USB)

Patch-set: 2