Update patch set 3
Patch Set 3: Code-Review+1

If this is the first time the user uses the superuser applicwtion he should read everything is written in the screen to understand what is going on. But after several accepts the user will touch accept without thinking. By this change i believe a user still has a lot of time to regret the approval and an experienced one can touch allow without second thought

Patch-set: 3