Update patch set 1
Patch Set 1:

I created the new types of exceptions because it is better for developers to track a bug or which "exceptional" code is run due to a bug. Also using Java 1.7 in Android we will can modify the source and convert the 3 different catches into one using the pipe ( | ) symbol between exception types.

Patch-set: 1