Update patch set 1
Patch Set 1: Code-Review+1

I think this is very important. Imagine the 2 following scenarios: SCENARIO A 1. doSystemInstall() is executed. 2. If in the try block a Throwable gets thrown we need the finally block to run extra code. SCENARIO B 1. doSystemInstall() is executed. 2. We have _error = false and at extractSu() a Throwable gets thrown. If we use finally the code is executed but the _error still is false! That's why I believe we need this commit.

Patch-set: 1