Update patch set 4
Patch Set 4:

I could remove the comments. But these are not written by me. I only commented (some of the lines) because I wanted to avoid any useless memory allocation with not used variables. I did not want to be the guy who removed these lines, but if it is necessary I could do this fix also.

Patch-set: 4