Update patch set 2
Patch Set 2:

Imagine that suddenly the DalvikVM is out memory and cannot create any more objects. It is not legal to catch an Error, so you should not create a catch block. Also an Error is when the device is out of space for saving files.
If finally is missing the code that is in finally block now will never be executed.
Also, if error = false and the finally exists you will inform the user that everything went fine even if an Error is thrown during the process.

Patch-set: 2