Update patch set 2
Patch Set 2:

Let's forget that we show an AlertDialog. We have source code, after the catch block, that needs the variable "error". We change the "error" at catch block. If we remove the changes from this commit the results will be the same, but the user will not be informed.
I think that these changes made the code (for me!) more readable and keep the user informed about what is happening at the background. As a personal point of view, I would not like not to be informed if action succeeded or not.

Patch-set: 2