Update patch set 2
Patch Set 2:

But in case "doSomething();" throws a subclass of Error the "System.out.println("bar");" will not executed.
If we want the application to behave like this, it is ok. I agree that this change is useless. But if we want the "System.out.println("bar");" to be executed in any case of Error we need finally. The question is:  Do we need to inform the user that something went wrong and this application is not responsible or we want to fail without any feeadback to the user?

Patch-set: 2