Update patch set 2
Patch Set 2:

As Error is an 'unrecoverable error' by definition, I think it's ok to just crash the app (and let the FC dialog pop up) in that case. The crash will happen anyway due to Error not being caught, it doesn't matter to additionally show an error message.

Patch-set: 2