Update patch set 1
Patch Set 1:

Well, this change isn't merged, so it can't be the culprit for issues you're seeing in the wild. AFAICT, the situation is like this:

- prior to http://review.cyanogenmod.org/#/c/50535/ : /storage/emulated was completely empty
- after #50535, but without this commit: /storage/emulated is populated except /storage/emulated/legacy and /storage/emulated/obb (because those are user-specific mounts, which we can't do in a globally running daemon). Additionally, /storage/emulated/legacy was unpopulated for the shell user, because the su daemon mount was globally visible.
- with this change, the mounts are as described above, except /storage/emulated/legacy being visible under the shell user again.

As explained in http://review.cyanogenmod.org/#/c/50826 for getting the legacy mount under su, it's required to have one subdaemon per UID (or at a minimum one per active user), which is what SuperSU does.

Patch-set: 1