Update patch set 1
Patch Set 1:

SuperSU may well break CifsManager and its kind, and that's unfortunate.

Personally I'm not convinced /storage/emulated needs to be correct, providing adb root shell semantics is sufficient for my purposes.  Plus, privileged programs can access /data/media directly anyways.

That said, if there are programs that do depend on the user-specific view of /storage/emulated, then I'd mount a tmpfs at /storage and mark that subtree as slave as we did in CM 10.1.  That said, I'm not sure how that impacts CTS, and the mount needs to be done in early-init, the patch for which was recently reverted in 10.2.

Patch-set: 1