Update patch set 3
Patch Set 3:

Minor nitpick: main() typically returns a value in the range of 0-255, because that is the output from WEXITSTATUS().  Here, main() returns -1 on failure, which will translate to WEXITSTATUS 255.

Is there a separate gerrit entry for sanitizing BOOTCLASSPATH?

Note that CLASSPATH is set to a hardcoded value from the /system/bin/am script, so it wasn't really a security risk in this case (at least on the ROMs I looked at).

Patch-set: 3