Update patch set 3
Patch Set 3:

For example. In this package 'developer_warning ' string: If I replace ? with a simply carriage return (enter key)  through Transifex, when I download the cm_translation file, this string doesn't show a correct newline character (\n). Only shows this string in three lines under xml code, but only one line is shown on device.

In this case, a correct translation for these strings is replace ? to \n. But be careful, for string-arrays, the <item> tag must be seperated with enter key for each one.

For all Spanish translations have been fixed.

Patch-set: 3