Update patch set 1
Patch Set 1:

Is this change cause of my problem: after flashing nightly later than this change, when an app requires Superuser privilege, the pop up appears, the 'Allow' button is greyed for a couple seconds and displays a countdown which when it reaches zero causes the 'Allow' button to appear to be clickable.  But touching it DOES NOTHING. It is impossible to [touch/click] ALLOW. Can click DENY, or at timeout, the superuser request gets denied. 
Nexus 5. Tried with nightly 2014-0726 and 2014-0725

Patch-set: 1