Update patch set 1
Patch Set 1:

Why do you need it? The main functionality of the app can already be done with the extended auto-brightness controls in CM. Lux will be just as problematic when installing apps, it will fail in the same way. That's due to an important security feature. Without it, malicious apps could overlay parts of the install or superuser dialogs, making it appears as though an innocent-looking app is being installed or one of your normal root apps is requesting root. It could also simply make it appear as though the allow and deny buttons are flipped, making you allow something you want to deny.

Patch-set: 1