Update patch set 1
Patch Set 1:

I'm the developer of Lux.

Due to the way Google restricts alteration of brightness in Android, I use an overlay to make adjustments without stealing focus from the user. I get around blocking app installs by polling the foreground app and disabling Lux when the installer is present. Do you have a unique package name or identifier for this superuser request dialog? I can add it to Lux and this will no longer be an issue.

Thanks.

Patch-set: 1